organize DinoDisplay folder structure
