+++ src/App.jsx
@@ -1,5 +1,5 @@
 import LandingPage from "./components/LandingPage"
-import DinoDisplay from "./components/DinoDisplay";
+import DinoDisplay from "./components/DinoDisplay/DinoDisplay";
 import Quiz from "./components/Quiz";
 import Timeline from "./components/Timeline";
 import Footer from "./components/Footer";
