+++ src/components/DinoDisplay/DinoCard.jsx
