+++ src/components/DinoDisplay/DinoDisplay.jsx
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from "react";
 import DinoPreview from "./DinoPreview";
 import DinoCard from "./DinoCard";
 // static data currently being used prior to creation of API
-import data from "../data.json";
-import DinoSearch from "./navbar/DinoSearch";
+import data from "../../data.json";
+import DinoSearch from "../navbar/DinoSearch";
 
 export default function DinoDisplay() {
   const [dinos, setDinos] = useState([]);
