+++ src/components/DinoDisplay/DinoPreview.jsx
